Changed application name to make this a separate project By now it is clear that my changes to multiload will never be merged back to nandhp. Changes are too many, too broad and too deep. Also, I contacted the author, but months later, he did not reply to me. Moreover, I'm planning to continue development of this plugin for a long time, and maybe I don't want to discuss every pull request I will make...
It is now time to make this a separate project. I changed application name, object names, GitHub
address and everything else to "multiload-ng" (hope I didn't forget to rename anything).
This name was chosen to emphasize the separation from the "previous" generation of multiload plugin.
Differences between original multiload and multiload-ng will increase over time, with enhancements
on compatibility, ahestetics and functionality. Future plans/to do list/wish list are all in a text
file that I was keeping for months. I will clean up and publish it soon as MarkDown document.

Important note: I don't mean to be in any way in competition with original plugin and its authors.
That's also why the object names were changed too: this allow both plugins to be installed
and used at the same time.