Add control to choose between two methods of counting used memory Basically user can choose whether memory count should be *procps-compliant* (current behavior) or not.
Being *procps-compliant* means that used/cached memory count is done
exactly the way it's done by `free` or `top`, that is, counting also
*Slab cache* into cached memory. Non *procps-compliant* behavior is
to not count *Slab cache* into cached memory, but into used. This is
the way it's done by *htop* and some graphical task managers (like *xfce4-taskmanager*)

This should answer once for all to all reports like #10 and #21.

No UI yet, this setting has to be changed manually.